fix: ensure consistent order of icons
This fix ensures that the icons in the manifest and Contents.json are
written in a deterministic order. This resolves issues where re-running
the tool on an unchanged set of icons could produce an unwanted change
to the local files, simply because the icon list was not in the same
order.

Fixes #67.